Rename path.hpp.
It conflicts too easily with subprojects (conflicts with Fossilize's path.hpp).
