+++ filesystem/android/android.cpp
@@ -21,7 +21,7 @@
  */
 
 #include "android.hpp"
-#include "path.hpp"
+#include "path_utils.hpp"
 #include "logging.hpp"
 #include <algorithm>
 #include <stdexcept>
