+++ filesystem/filesystem.cpp
@@ -21,7 +21,7 @@
  */
 
 #include "filesystem.hpp"
-#include "path.hpp"
+#include "path_utils.hpp"
 #include "logging.hpp"
 #include "os_filesystem.hpp"
 #include "string_helpers.hpp"
