+++ filesystem/volatile_source.hpp
@@ -22,7 +22,7 @@
 
 #pragma once
 
-#include "path.hpp"
+#include "path_utils.hpp"
 #include "filesystem.hpp"
 #include "intrusive.hpp"
 #include "logging.hpp"
