+++ filesystem/windows/os_filesystem.hpp
@@ -21,7 +21,7 @@
  */
 
 #pragma once
-#include "../filesystem.hpp"
+#include "filesystem.hpp"
 #include <unordered_map>
 #define WIN32_LEAN_AND_MEAN
 #include <windows.h>
