+++ path/path_utils.cpp
@@ -20,7 +20,7 @@
  * SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 
-#include "path.hpp"
+#include "path_utils.hpp"
 #include "string_helpers.hpp"
 #include <algorithm>
